requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==41.0.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.1.9
django-allauth==0.54.0
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.20.0
google-cloud-speech==2.20.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
gTTS==2.3.2
idna==3.4
image==1.5.33
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
mysqlclient==2.2.0
nest-asyncio==1.5.6
oauthlib==3.2.2
openai==0.27.8
optional-django==0.1.0
packaging==22.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.7.0
pymongo==3.12.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==305
pyzmq==24.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.2.4
stack-data==0.6.2
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.5
yarl==1.9.2
